using std randint

+++ src/geneticTask/lib/words.cpp
@@ -9,7 +9,7 @@ int32_t prefixes::get_template(const u_int8_t word0) {
     words[0] = word0;
     for (int i = 0; i < 3; i++) {
         prefTemplate curTemplate = templates[word0][i];
-        words[i + 1] = curTemplate.startPrefix - minDiff * randint(0, curTemplate.availableSteps); 
+        words[i + 1] = curTemplate.startPrefix - minDiff * randint(0, (int)curTemplate.availableSteps); 
     }
     
     if (word0 == word0::ASSIGN) {
